requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
streamlit>=1.22.0
requests>=2.28.0
python-dotenv>=0.21.0
together>=0.2.0